python-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
anthropic>=0.25.0
pydantic==2.4.2
python-multipart==0.0.6
qdrant-client==1.14.2
huggingface_hub==0.16.4
transformers==4.30.2
sentence-transformers==2.2.2
torch==2.0.1
pillow==10.0.0
opencv-python==4.8.0.74
numpy==1.24.3
requests==2.31.0